server/requirements.txt
flask
numpy
imageio
Pillow
nltk
h5py
imageio
torchvision
torch
